# specs/grc-claw-openapi.yaml
openapi: 3.1.0
info:
  title: GRC_Claw API
  description: |
    GRC_Claw — open-source governance, risk, and compliance platform for agentic AI.
    
    This API provides four complementary interfaces:
    - **REST** — CRUD operations, search, and reporting
    - **GraphQL** — Complex nested queries for dashboards and analytics
    - **gRPC** — High-performance enforcement decisions and streaming
    - **Webhooks** — Event-driven integrations
    
    ## Authentication
    
    All API requests require authentication via one of:
    - **OAuth 2.1 / OIDC** — Bearer JWT token (RS256)
    - **API Key** — `Authorization: Bearer grc_live_...` or `grc_test_...`
    - **mTLS (SPIFFE)** — X.509 SVID for agent-to-PEP communication
    
    ## Rate Limiting
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit` — Request limit per window
    - `X-RateLimit-Remaining` — Remaining requests in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    
    ## Error Format
    
    All errors follow RFC 7807 (Problem Details for HTTP APIs).
    
    ## Versioning
    
    Current version: `v1.0`. Version is specified in the URL path.
    Deprecation notices are sent via webhook 6 months before sunset.
  version: 1.0.0
  contact:
    name: GRC_Claw Architecture Team
    url: https://api.grc-claw.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.grc-claw.io
...
paths:
...
  /v1.0/enforcement/decide:
    post:
      operationId: enforcementDecide
      summary: Request enforcement decision
      description: Evaluates agent action against policies and returns a decision.
      tags: [Enforcement]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnforcementRequest'
      responses:
        '200':
          description: Enforcement decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnforcementDecision'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/enforcement/decide-batch:
    post:
      operationId: enforcementDecideBatch
      summary: Batch enforcement decisions
      description: Evaluates multiple agent actions in a single request.
      tags: [Enforcement]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decisions]
              properties:
                decisions:
                  type: array
                  items:
                    $ref: '#/components/schemas/EnforcementRequest'
      responses:
        '200':
          description: Batch results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/EnforcementDecision'
                  summary:
                    $ref: '#/components/schemas/BatchSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/enforcement/decisions/{decision_id}:
    get:
      operationId: getDecision
      summary: Get decision
      description: Returns full decision certificate with evidence.
      tags: [Enforcement]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Decision details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnforcementDecision'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/enforcement/decisions:
    get:
      operationId: listDecisions
      summary: List decisions
      description: Returns paginated list of enforcement decisions.
      tags: [Enforcement]
      parameters:
        - name: agent_id
          in: query
          schema: { type: string }
        - name: policy_id
          in: query
          schema: { type: string }
        - name: verdict
          in: query
          schema:
            type: string
            enum: [ALLOW, ALLOW_WITH_REDACTION, REQUIRE_APPROVAL, DENY, QUARANTINE]
        - name: date_from
          in: query
          schema: { type: string, format: date-time }
        - name: date_to
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated decisions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1.0/compliance/frameworks:
    get:
      operationId: listFrameworks
      summary: List compliance frameworks
      description: Returns all supported compliance frameworks.
      tags: [Compliance]
      responses:
        '200':
          description: List of frameworks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplianceFramework'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/compliance/frameworks/{framework_id}/controls:
    get:
      operationId: listControls
      summary: List controls
      description: Returns controls for a specific framework.
      tags: [Compliance]
      parameters:
        - name: framework_id
          in: path
          required: true
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: target_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of controls
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplianceControl'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/compliance/posture:
    get:
      operationId: getCompliancePosture
      summary: Get compliance posture
      description: Returns compliance posture for a framework and target.
      tags: [Compliance]
      parameters:
        - name: framework
          in: query
          required: true
          schema: { type: string }
        - name: target_id
          in: query
          required: true
          schema: { type: string }
        - name: target_type
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Compliance posture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompliancePosture'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1.0/compliance/reports:
    post:
      operationId: generateComplianceReport
      summary: Generate compliance report
      description: Initiates async compliance report generation.
      tags: [Compliance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceReportRequest'
      responses:
        '202':
          description: Report generation initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  status:
                    type: string
                    enum: [processing]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/compliance/crosswalk:
    get:
      operationId: crosswalkQuery
      summary: Crosswalk query
      description: Returns cross-framework control mappings.
      tags: [Compliance]
      parameters:
        - name: control_id
          in: query
          schema: { type: string }
        - name: framework
          in: query
          schema: { type: string }
        - name: target_framework
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Crosswalk mappings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrosswalkResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key in format: Bearer grc_live_... or Bearer grc_test_...'
    mtlsAuth:
      type: mutualTLS
      description: mTLS with SPIFFE SVID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        trace_id:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    EnforcementRequest:
      type: object
      required: [agent_id, action, resource]
      properties:
        agent_id: { type: string }
        action: { type: string }
        resource: { type: string }
        context:
          type: object
          additionalProperties: true
        policy_ids:
          type: array
          items: { type: string }
        include_evidence: { type: boolean }
    EnforcementDecision:
      type: object
      properties:
        decision_id: { type: string }
        verdict:
          type: string
          enum: [ALLOW, ALLOW_WITH_REDACTION, REQUIRE_APPROVAL, DENY, QUARANTINE]
        policy_id: { type: string }
        policy_version: { type: string }
        agent_id: { type: string }
        action: { type: string }
        resource: { type: string }
        context:
          type: object
          additionalProperties: true
        evidence_hash: { type: string }
        timestamp:
          type: string
          format: date-time
        ttl: { type: integer }
        signature: { type: string }
        matched_rules:
          type: array
          items: { type: string }
        evaluation_time_ms: { type: number }
        reason: { type: string }
    BatchSummary:
      type: object
      properties:
        total: { type: integer }
        allowed: { type: integer }
        denied: { type: integer }
        require_approval: { type: integer }
        quarantined: { type: integer }
        avg_evaluation_time_ms: { type: number }
    DecisionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EnforcementDecision'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    ComplianceFramework:
      type: object
      properties:
        id: { type: string }
        framework_key: { type: string }
        name: { type: string }
        version: { type: string }
        description: { type: string }
        authority: { type: string }
        effective_date:
          type: string
          format: date-time
        control_count: { type: integer }
    ComplianceControl:
      type: object
      properties:
        id: { type: string }
        framework_id: { type: string }
        control_key: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        guidance: { type: string }
    CompliancePosture:
      type: object
      properties:
        framework: { type: string }
        target_id: { type: string }
        target_type: { type: string }
        controls_assessed: { type: integer }
        controls_compliant: { type: integer }
        controls_non_compliant: { type: integer }
        controls_not_assessed: { type: integer }
        compliance_score: { type: number }
        gaps:
          type: array
          items:
            type: object
            properties:
              control_id: { type: string }
              control_title: { type: string }
              status: { type: string }
              severity: { type: string }
              evidence_count: { type: integer }
              last_assessed:
                type: string
                format: date-time
        trend:
          type: object
          properties:
            direction: { type: string }
            change: { type: string }
            period: { type: string }
...
    ComplianceReportRequest:
      type: object
      required: [framework, time_range, format]
      properties:
        framework: { type: string }
        time_range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        format:
          type: string
          enum: [json, pdf, csv, xml]
        include_evidence: { type: boolean }
        include_gaps: { type: boolean }
    CrosswalkResult:
      type: object
      properties:
        grc_control_id: { type: string }
        mappings:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...